#364: divided actions + comment
