+++ webapp/app/containers/ForgottenPassword/Form/Email/actions.js
@@ -1,5 +1,5 @@
 //
-// ForgottenPassword actions
+// ForgottenPassword email form actions
 //
 // To add a new Action :
 // 1) Import you constant
@@ -8,10 +8,6 @@
 //        return { type: YOUR_ACTION_CONSTANT, var: var }
 //    }
 //
-
-import request from 'utils/request';
-import { push } from 'react-router-redux';
-import { browserHistory } from 'react-router';
 
 import {
   FORGOTTEN_PASSWORD,
@@ -29,25 +25,5 @@ export function forgottenPassword(res) {
   return {
     type: FORGOTTEN_PASSWORD,
     res,
-  };
-}
-
-export function forgottenPasswordRequest(email) {
-  return function returnForgottenPasswordRequest(dispatch) {
-    return request
-      .post('/api/public/user/recover_password')
-      .type('form')
-      .send({ email })
-      .end((err, res) => {
-        dispatch(forgottenPassword(res.body));
-
-        if (err && res.statusCode === 401) {
-          browserHistory.push('/login');
-        }
-
-        if (res.body.success) {
-          dispatch(push('/login'));
-        }
-      });
   };
 }
